% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-action-applicability.R
\name{add_applicability}
\alias{add_applicability}
\alias{remove_applicability}
\alias{update_applicability}
\title{Add an applicability domain to a workflow}
\usage{
add_applicability(x, method, predictors = NULL, ...)

remove_applicability(x)

update_applicability(x, method, predictors = NULL, ...)
}
\arguments{
\item{x}{A workflow}

\item{method}{The type of applicability domain model to use. Current values
are: "hat_values", "isolation", "pca", and "similarity".}

\item{predictors}{Tidyselect expressions specifying the original predictors
of the  model. See \link[tidyselect:language]{tidyselect::select_helpers} for the full range of
possible ways  to specify terms. The default is to use all predictors given
to the workflow.}

\item{...}{Options to pass to the applicability domain model.}
}
\value{
\code{x}, updated with either a new or removed applicability postprocessor.
}
\description{
\itemize{
\item \code{add_applicability()} specifies the type of applicability domain model that
should be used.
\item \code{remove_applicability()} removes the applicability domain model as well as
any downstream objects that might get created after the applicability
domain is used for post-processing.
\item \code{update_applicability()} first removes the applicability domain model, then
replaces it with the new one.
}
}
\examples{
library(applicable)
}
