% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-action-case-weights.R
\name{add_case_weights}
\alias{add_case_weights}
\alias{remove_case_weights}
\alias{update_case_weights}
\title{Add case weights to a workflow}
\usage{
add_case_weights(x, col)

remove_case_weights(x)

update_case_weights(x, col)
}
\arguments{
\item{x}{A workflow}

\item{col}{A single unquoted column name specifying the case weights for
the model.}
}
\description{
\itemize{
\item \code{add_case_weights()} specifies the column that will be interpreted as
case weights in the model. This column must be present in the \code{data}
supplied to \link[=fit.workflow]{fit()}.
\item \code{remove_case_weights()} removes the case weights. Additionally, if the
model has already been fit, then the fit is removed.
\item \code{update_case_weights()} first removes the case weights, then replaces them
with the new ones.
}
}
\details{
For formula and variable preprocessors, the case weights \code{col} is removed
from the data before the preprocessor is evaluated. This allows you to use
formulas like \code{y ~ .} or tidyselection like \code{everything()} without fear of
accidentally selecting the case weights column.

For recipe preprocessors, the case weights \code{col} is not removed and is
passed along to the recipe. Typically, your recipe will include steps that
can utilize case weights.
}
\examples{
library(parsnip)
library(magrittr)

spec <- linear_reg() \%>\%
  set_engine("lm")

wf <- workflow() \%>\%
  add_case_weights(gear) \%>\%
  add_formula(mpg ~ .) \%>\%
  add_model(spec)

wf <- fit(wf, mtcars)

# Notice that the case weights (gear) aren't included in the predictors
extract_mold(wf)$predictors

# Strip them out of the workflow, which also resets the model
remove_case_weights(wf)
}
