% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-action-container.R
\name{add_container}
\alias{add_container}
\alias{remove_container}
\alias{update_container}
\title{Add a container to a workflow}
\usage{
add_container(x, container, ...)

remove_container(x)

update_container(x, container, ...)
}
\arguments{
\item{x}{A workflow}

\item{container}{A container created using \code{\link[container:container]{container::container()}}. The container
should not have been trained already with \code{\link[container:reexports]{container::fit()}}; workflows
will handle training internally.}

\item{...}{Not used.}
}
\value{
\code{x}, updated with either a new or removed container postprocessor.
}
\description{
\itemize{
\item \code{add_container()} specifies post-processing steps to apply through the
usage of a container.
\item \code{remove_container()} removes the container as well as any downstream objects
that might get created after the container is used for post-processing, such as
the fitted container.
\item \code{update_container()} first removes the container, then replaces the postvious
container with the new one.
}
}
\examples{
library(container)
library(magrittr)

container <- container() \%>\%
  adjust_probability_threshold(.1)

workflow <- workflow() \%>\%
  add_container(container)

workflow

remove_container(workflow)

update_container(workflow, container() \%>\% adjust_probability_threshold(.2))
}
