% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-action-model.R
\name{add_model}
\alias{add_model}
\title{Add a model to a workflow}
\usage{
add_model(x, spec, formula = NULL)
}
\arguments{
\item{x}{A workflow.}

\item{spec}{A parsnip model specification.}

\item{formula}{An optional formula override to specify the terms of the
model. Typically, the terms are extracted from the formula or recipe
preprocessing methods. However, some models (like survival and bayesian
models) use the formula not to preprocess, but to specify the structure
of the model. In those cases, a formula specifying the model structure
must be passed unchanged into the model call itself. This argument is
used for those purposes.}
}
\description{
\code{add_model()} adds a parsnip model to the workflow.
}
\details{
\code{add_model()} is a required step to construct a minimal workflow.
}
