% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-action-recipe.R
\name{add_recipe}
\alias{add_recipe}
\title{Add a recipe to a workflow}
\usage{
add_recipe(x, recipe)
}
\arguments{
\item{x}{A workflow.}

\item{recipe}{A recipe created using \code{\link[recipes:recipe]{recipes::recipe()}}.}
}
\description{
\code{add_recipe()} specifies the terms of the model and any preprocessing that
is required through the usage of a recipe.
}
\details{
To fit a workflow, one of \code{add_formula()} or \code{add_recipe()} \emph{must} be
specified, but not both.
}
