% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-action-tailor.R
\name{add_tailor}
\alias{add_tailor}
\alias{remove_tailor}
\alias{update_tailor}
\title{Add a tailor to a workflow}
\usage{
add_tailor(x, tailor, ...)

remove_tailor(x)

update_tailor(x, tailor, ...)
}
\arguments{
\item{x}{A workflow}

\item{tailor}{A tailor created using \code{\link[tailor:tailor]{tailor::tailor()}}. The tailor
should not have been trained already with \code{\link[tailor:reexports]{tailor::fit()}}; workflows
will handle training internally.}

\item{...}{Not used.}
}
\value{
\code{x}, updated with either a new or removed tailor postprocessor.
}
\description{
\itemize{
\item \code{add_tailor()} specifies post-processing steps to apply through the
usage of a tailor.
\item \code{remove_tailor()} removes the tailor as well as any downstream objects
that might get created after the tailor is used for post-processing, such as
the fitted tailor.
\item \code{update_tailor()} first removes the tailor, then replaces the postvious
tailor with the new one.
}
}
\examples{
library(tailor)
library(magrittr)

tailor <- tailor("binary")
tailor_1 <- adjust_probability_threshold(tailor, .1)

workflow <- workflow() \%>\%
  add_tailor(tailor_1)

workflow

remove_tailor(workflow)

update_tailor(workflow, adjust_probability_threshold(tailor, .2))
}
