% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-manipulation-methods.R
\name{bind}
\alias{bind}
\title{Combine two pipelines into one}
\usage{
bind(x, ...)
}
\arguments{
\item{x}{Object of class "pipeline".}

\item{...}{Other arguments.}

\item{elements}{Elements of \code{data} attribute to combine.}

\item{by}{Not used.}
}
\value{
A list of class "pipeline".
}
\description{
Combine two pipelines into a single pipeline. This pipeline has attribiutes from the
first pipeline of the input, except the \code{elements}, which are row-bound together.
}

