% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{call_factory}
\alias{call_factory}
\title{Generate calls of function applied to arguments}
\usage{
call_factory(f, ...)
}
\arguments{
\item{...}{Objects to apply f to.}

\item{x}{Function.}
}
\value{
A call.
}
\description{
Generate calls of function applied to arguments.
}

