% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{ctrl_workflow}
\alias{ctrl_workflow}
\title{Control object for a workflow}
\usage{
ctrl_workflow(ctrl_parsnip = fit_control())
}
\arguments{
\item{ctrl_parsnip}{A parsnip control object.}
}
\description{
\code{ctrl_workflow()} holds the control parameters for a workflow.
}
