% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-construction-methods.R
\name{delay}
\alias{delay}
\title{Add a step to a workflow or pipeline}
\usage{
delay(x, ...)
}
\arguments{
\item{x}{Object of class "workflow" or "pipeline".}

\item{...}{Other arguments.}

\item{method}{Method to execute.}
}
\value{
If \code{x} is a "workflow", then a list of class "workflow" with a new "target" for \code{method} in each "pipeline". If
       \code{x} is a "pipeline", then a pipeline with a new "target" for \code{method}.
}
\description{
Add a step to a "workflow" or "pipeline" object where \code{method} will be executed.
}

