% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-execution-methods.R
\name{execute}
\alias{execute}
\title{Execute a Workflow or Pipeline}
\usage{
execute(x, ...)
}
\arguments{
\item{x}{List of class "workflow" or "pipeline". If workflow, builds the targets in the Pipelines; if Pipeline, builds the targets.}

\item{...}{Other arguments passed to execute_distributed.}

\item{locality}{If "locality", execute on local machine. If "sge", spin up one SGE job per parallel pipeline and wait for results;
requires a GridEngine cluster manager.}

\item{parallel}{Logical. If \code{TRUE}, use all available cores for (hopefully) a performance gain.}

\item{clobber}{Logical. Execute from the very beginning (\code{TRUE}) or from last-built Target (\code{FALSE}).}
}
\value{
List of class "workflow" or "pipeline" with targets built.
}
\description{
Execute the Targets in Pipeleines.
}
\seealso{
\link{execute_distributed}
}

