% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-execution-methods.R
\name{execute_distributed}
\alias{execute_distributed}
\title{Execute a Pipeline on SGE}
\usage{
execute_distributed(x, ...)
}
\arguments{
\item{x}{List of class "workflow" or "pipeline". If workflow, builds the targets in the Pipelines; if Pipeline, builds the targets.}

\item{...}{Other arguments.}

\item{mode}{Not used.}

\item{submit_command}{Character. Command used for job submission; usually "qsub" (the default).}

\item{submit_script}{Character. Path to a SGE submission script template.}

\item{clobber}{Logical. Execute from the very beginning (\code{TRUE}) or from last-built Target (\code{FALSE}).}
}
\value{
List of class "SGEjob" containing submission parameters.
}
\description{
Execute the Targets in Pipeleines using a GridEngine style cluster.
}

