% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{get_unique_values}
\alias{get_unique_values}
\alias{get_unique_values.default}
\alias{get_unique_values.data.frame}
\title{Get unique values}
\usage{
get_unique_values(x, ...)

\method{get_unique_values}{default}(x, ...)

\method{get_unique_values}{data.frame}(x, column, sort_result = TRUE,
  ...)
}
\arguments{
\item{x}{Object of class "data.frame"".}

\item{...}{Other arguments.}

\item{column}{Character. Element of \code{x} to find unique values for.}

\item{sort_result}{Logical. Sort the unique values before returning?}
}
\value{
Vector of same class as \code{column}
}
\description{
Get unique values of an element of an object.
}
