% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Data retrieval}
\usage{
getData(vignLabel)
}
\arguments{
\item{vignLabel}{A short label specifying which vignette to get the data for. See Details.}
}
\description{
Downloads relevant data for individual workflow vignettes, usually from Synapse
}
\details{
All data is downloaded into a local data/ directory

Use the following values for \strong{vignLabel} to designate which dataset to load:
\describe{
  \item{"dge-wrangle"}{Example output from bcbio that is used by the DGE wrangling vignette}
  \item{"dge-pca"}{Data for the vignette that demonstrates how to apply Principal Components
                 Analysis to a counts matrix from a DGE experiment}
}
}
