% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-construction-methods.R
\name{is_locked}
\alias{is_locked}
\title{Check if Pipeline is locked by another process}
\usage{
is_locked(x, ...)
}
\arguments{
\item{x}{List of class "pipeline".}

\item{...}{Other arguments.}
}
\value{
\code{TRUE} or \code{FALSE} depending on presence of Pipeline lockfile.
}
\description{
Rerturns \code{TRUE} or \code{FALSE} depending on presence of Pipeline lockfile as defined in the \code{lock_file}
attribute.
}

