% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge-management.R
\name{is_queued}
\alias{is_queued}
\title{Check if SGEjob still in queue}
\usage{
is_queued(x, ...)
}
\arguments{
\item{x}{List of class "SGEjob".}

\item{...}{Other arguments.}
}
\value{
\code{TRUE} if SGEjob still in queue, otherwise \code{FALSE}.
}
\description{
Check if SGEjob still in queue
}

