% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge-management.R
\name{is_running}
\alias{is_running}
\title{Check if SGEjob still running}
\usage{
is_running(x, ...)
}
\arguments{
\item{x}{List of class "SGEjob".}

\item{...}{Other arguments.}
}
\value{
\code{TRUE} if SGEjob still running, otherwise \code{FALSE}.
}
\description{
Check if SGEjob still running
}
