% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge-management.R
\name{is_successful}
\alias{is_successful}
\title{Check if SGEjob ended successfully}
\usage{
is_successful(x, ...)
}
\arguments{
\item{x}{List of class "SGEjob".}

\item{...}{Other arguments.}
}
\value{
\code{TRUE} if SGEjob ended successfully, otherwise \code{FALSE}.
}
\description{
Check if SGEjob ended successfully, as determined by existence of file at path \code{success_filename}.
}
