% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-construction-methods.R
\name{load_workflow}
\alias{load_workflow}
\title{Load Workflow from disk}
\usage{
load_workflow(filename, ...)
}
\arguments{
\item{filename}{Character. Path to serialized Workflow.}

\item{...}{Other arguments.}
}
\value{
A list of class "workflow".
}
\description{
Load serialized Workflow from \code{filename}.
}
\seealso{
\link{readRDS}
}
