% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{mclapply2}
\alias{mclapply2}
\title{Generate calls of function applied to arguments}
\usage{
mclapply2(X, FUN, ..., mc.preschedule = TRUE, mc.set.seed = TRUE,
  mc.silent = FALSE, mc.cores = get_cores(), mc.cleanup = TRUE,
  mc.allow.recursive = FALSE)
}
\arguments{
\item{...}{Objects to apply f to.}

\item{x}{Function.}
}
\value{
A call.
}
\description{
Generate calls of function applied to arguments.
}

