% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-object-definition.R
\name{newPipeline}
\alias{newPipeline}
\title{Make a new Pipeline object.}
\usage{
newPipeline(input_data, checkpoint = FALSE, working_directory = ".",
  ...)
}
\arguments{
\item{input_data}{Starting data}

\item{checkpoint}{Boolean. Checkpoint intermediate files? Default \code{FALSE}.}

\item{working_directory}{Character. Top-level directory for data processing. Default is current working directory.}

\item{...}{Other arguments}
}
\value{
list of class "pipeline"
}
\description{
Make a new Pipeline object.
}
