% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target-object-definition.R
\name{newTarget}
\alias{newTarget}
\title{Make a new Target object.}
\usage{
newTarget(method, input_file = NA, output_file = NA, ...)
}
\arguments{
\item{method}{Name. Method to apply.}

\item{input_file}{Not used.}

\item{output_file}{Not used.}

\item{...}{Other arguments passed to \code{call_factory()}.}
}
\value{
list of class "target"
}
\description{
Make a new Target object.
}
