% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-object-definition.R
\name{newWorkflow}
\alias{newWorkflow}
\title{Make a new Worklow object}
\usage{
newWorkflow(input_data, working_directory = ".", checkpoint = FALSE,
  load_checkpoint = NULL, ...)
}
\arguments{
\item{input_data}{Starting data
working_directory Character. Top-level directory for data processing. Default is current working directory.
checkpoint Boolean. Checkpoint intermediate files? Default \code{FALSE}.
load_checkpoint Character. Start from named saved checkpoint. If \code{NULL} (default), don't.
... Other arguments}
}
\value{
list of class "workflow"
}
\description{
Make a new Worklow object
}

