% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-manipulation-methods.R
\name{scatter}
\alias{scatter}
\alias{scatter.default}
\alias{scatter.workflow}
\title{Split pipeline data into parallel chunks}
\usage{
scatter(x, ...)

\method{scatter}{default}(x, ...)

\method{scatter}{workflow}(x, elements, by, ...)
}
\arguments{
\item{x}{Object of class "workflow" or "pipeline".}

\item{...}{Other arguments.}

\item{elements}{Elements of \code{data} attribute to look for matches.}

\item{by}{Column of \code{elements} to chunk on. Each unique value contained in column \code{by} is split into a separate pipeline.}
}
\value{
If \code{x} is a "workflow", then a list of class "workflow" containing one "pipeline" for each value of \code{by}. If
       \code{x} is a "pipeline", then a list of "pipeline"s.
}
\description{
Split pipeline data into parallel chunks by some filtering condition. Each unique value contained in
column \code{by} in the \code{elements} of the \code{data} attribute of each "pipeline" is split into a separate pipeline.
All other attributes are copied.
}
