% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stackoverflow}
\alias{stackoverflow}
\title{Annual Stack Overflow Developer Survey Data}
\source{
Julia Silge, \emph{Supervised Machine Learning Case Studies in R},
\url(https://supervised-ml-course.netlify.com/chapter2)
}
\value{
\item{stackoverflow}{a tibble}
}
\description{
Annual Stack Overflow Developer Survey Data
}
\details{
These data are a collection of 5,594 data points collected on
developers. These data could be used to try to predict who works remotely
(as used in the source listed below).
}
\examples{
data(stackoverflow)
}
\keyword{datasets}
