% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge-management.R
\name{status}
\alias{status}
\title{Check on SGEjob status}
\usage{
status(x, ...)
}
\arguments{
\item{x}{List of class "SGEjob".}

\item{...}{Other arguments.}
}
\value{
Output of \code{qstat} for \code{jobid} of SGEjob.
}
\description{
Check on SGEjob status by getting output of \code{qstat}.
}
