% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butcher.R
\name{workflow-butcher}
\alias{workflow-butcher}
\alias{axe_call.workflow}
\alias{axe_ctrl.workflow}
\alias{axe_data.workflow}
\alias{axe_env.workflow}
\alias{axe_fitted.workflow}
\title{Butcher methods for a workflow}
\usage{
axe_call.workflow(x, verbose = FALSE, ...)

axe_ctrl.workflow(x, verbose = FALSE, ...)

axe_data.workflow(x, verbose = FALSE, ...)

axe_env.workflow(x, verbose = FALSE, ...)

axe_fitted.workflow(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{A workflow.}

\item{verbose}{Should information be printed about how much memory is freed
from butchering?}

\item{...}{Extra arguments possibly used by underlying methods.}
}
\description{
These methods allow you to use the butcher package to reduce the size of
a workflow. After calling \code{butcher::butcher()} on a workflow, the only
guarantee is that you will still be able to \code{predict()} from that workflow.
Other functions may not work as expected.
}
