% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow}
\alias{workflow}
\title{Create a workflow}
\usage{
workflow()
}
\value{
A new \code{workflow} object.
}
\description{
A \code{workflow} is a container object that aggregates information required to
fit and predict from a model. This information might be a recipe used in
preprocessing, specified through \code{\link[=add_recipe]{add_recipe()}}, or the model specification
to fit, specified through \code{\link[=add_model]{add_model()}}.
}
\examples{
library(recipes)

rec <- recipe(mpg ~ cyl, mtcars)
rec <- step_log(rec, cyl)

wrk <- workflow()
wrk <- add_recipe(wrk, rec)

}
