% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflowier-package.R
\docType{package}
\name{workflowier-package}
\alias{workflowier}
\alias{workflowier-package}
\title{workflowier: Workflows}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/workflowier}
  \item Report bugs at \url{https://github.com/DavisVaughan/workflowier/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

}
\keyword{internal}
