% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows-package.R
\docType{package}
\name{workflows-package}
\alias{workflows}
\alias{workflows-package}
\title{workflows: Workflows}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/workflows}
  \item Report bugs at \url{https://github.com/DavisVaughan/workflows/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

}
\keyword{internal}
