% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-construction-methods.R
\name{write.workflow}
\alias{write.workflow}
\title{Write Workflow to disk.}
\usage{
write.workflow(x, filename, ...)
}
\arguments{
\item{x}{Object of class Workflow.}

\item{filename}{Character. Filename to serialize \code{x} under.}

\item{...}{Other arguments.}
}
\value{
The input Workflow, invisibly.
}
\description{
Save serialized Workflow \code{x} as \code{filename}
}
\seealso{
\link{\code{saveRDS}}
}

