% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-construction-methods.R
\name{write_workflow}
\alias{write_workflow}
\title{Write Workflow to disk.}
\usage{
write_workflow(x, filename, cleanup = TRUE, ...)
}
\arguments{
\item{x}{Object of class Workflow.}

\item{filename}{Character. Filename to serialize \code{x} under.}

\item{cleanup}{Logical. Delete intermediate pipeline checkpoints after saving?}

\item{...}{Other arguments.}
}
\value{
The input Workflow, invisibly.
}
\description{
Save serialized Workflow \code{x} as \code{filename}
}
\seealso{
\link{saveRDS}
}

