% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-action-case-weights.R
\name{add_case_weights}
\alias{add_case_weights}
\alias{remove_case_weights}
\alias{update_case_weights}
\title{Add case weights to a workflow}
\usage{
add_case_weights(x, col, ..., remove = TRUE)

remove_case_weights(x)

update_case_weights(x, col, ..., remove = TRUE)
}
\arguments{
\item{x}{A workflow}

\item{col}{A single unquoted column name specifying the case weights for
the model.}

\item{...}{These dots are for future extensions and must be empty.}

\item{remove}{A single logical specifying whether or not to remove the case
weights column from the \code{data} supplied to \code{fit()} before the preprocessor
is applied.

For formula preprocessors, variables preprocessors, and for recipe
preprocessors that don't use case weights, this should always be \code{TRUE}.
This ensures that formulas such as \code{y ~ .} won't pick up the case weights
column as a predictor.

For recipe preprocessors that use case weights, if the case weights column
used in the recipe is the exact same column as \code{col}, then set \code{remove = FALSE} to retain the case weights column in \code{data} and pass it on to the
preprocessor.}
}
\description{
\itemize{
\item \code{add_case_weights()} specifies the column that will be interpreted as
case weights in the model. This column must be present in the \code{data}
supplied to \link[=fit.workflow]{fit()}.
\item \code{remove_case_weights()} removes the case weights. Additionally, if the
model has already been fit, then the fit is removed.
\item \code{update_case_weights()} first removes the case weights, then replaces them
with the new ones.
}
}
\examples{
library(parsnip)
library(magrittr)

spec <- linear_reg() \%>\%
  set_engine("lm")

wf <- workflow() \%>\%
  add_case_weights(gear) \%>\%
  add_formula(mpg ~ .) \%>\%
  add_model(spec)

wf <- fit(wf, mtcars)

# Notice that the case weights (gear) aren't included in the predictors
extract_mold(wf)$predictors

# Strip them out of the workflow, which also resets the model
remove_case_weights(wf)
}
