% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-action-formula.R
\name{add_formula}
\alias{add_formula}
\title{Add formula terms to a workflow}
\usage{
add_formula(x, formula)
}
\arguments{
\item{x}{A workflow}

\item{formula}{A formula specifying the terms of the model. It is advised to
not do preprocessing in the formula, and instead use a recipe if that is
required.}
}
\description{
\code{add_formula()} specifies the terms of the model through the usage of a
formula.
}
\details{
To fit a workflow, one of \code{add_formula()} or \code{add_recipe()} \emph{must} be
specified, but not both.
}
\examples{
workflow <- workflow()
workflow <- add_formula(workflow, mpg ~ cyl)
workflow
}
