% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-manipulation-methods.R
\name{bind}
\alias{bind}
\alias{bind.default}
\alias{bind.pipeline}
\title{Combine two pipelines into one}
\usage{
bind(x1, ...)

\method{bind}{default}(x1, ...)

\method{bind}{pipeline}(x1, x2, elements, ...)
}
\arguments{
\item{x1}{Object of class "pipeline".}

\item{...}{Other arguments.}

\item{x2}{Object of class "pipeline".}

\item{elements}{Elements of \code{data} attribute to combine.}
}
\value{
A list of class "pipeline".
}
\description{
Combine two pipelines into a single pipeline. This pipeline has attribiutes from the
first pipeline of the input, except the \code{elements}, which are row-bound together.
}

