% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target-object-methods.R, R/workflow-object-methods.R
\name{build}
\alias{build}
\title{Execute a Target object according to its call and provided input data.}
\usage{
build(x, ...)

build(x, ...)
}
\arguments{
\item{x}{Object of class "target".
data Apply the target's method to this data.
... Other arguments.}

\item{x}{Object of class "target".
data Apply the target's method to this data.
... Other arguments.}
}
\value{
The result of applying the target's call to the \code{data}.

The result of applying the target's call to the \code{data}.
}
\description{
Execute a Target object according to its call and provided input data.

Execute a Target object according to its call and provided input data.
}

