% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-construction-methods.R
\name{checkpoint}
\alias{checkpoint}
\title{Checkpoint Pipeline to disk}
\usage{
checkpoint(x, ...)
}
\arguments{
\item{x}{Object of class "workflow" or "pipeline".}

\item{...}{Other arguments.}
}
\value{
A list of class "pipeline" invisibly.
}
\description{
Serializes Pipeline object under its \code{checkpoint_filename} attribute.
}

