% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{control_workflow}
\alias{control_workflow}
\title{Control object for a workflow}
\usage{
control_workflow(control_parsnip = fit_control())
}
\arguments{
\item{control_parsnip}{A parsnip control object.}
}
\description{
\code{control_workflow()} holds the control parameters for a workflow.
}
