% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{mclapply2}
\alias{mclapply2}
\title{#@title Generate calls of function applied to arguments
#@description Generate calls of function applied to arguments.
#@param X Function.
#@param ... Objects to apply f to.
#@return A call.}
\usage{
mclapply2(X, FUN, ..., mc.preschedule = TRUE, mc.set.seed = TRUE,
  mc.silent = FALSE, mc.cores = get_cores(), mc.cleanup = TRUE,
  mc.allow.recursive = FALSE)
}
\description{
#@title Generate calls of function applied to arguments
#@description Generate calls of function applied to arguments.
#@param X Function.
#@param ... Objects to apply f to.
#@return A call.
}
