% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{required_pkgs.workflow}
\alias{required_pkgs.workflow}
\title{Determine required packages for a workflow}
\usage{
\method{required_pkgs}{workflow}(x, infra = TRUE, ...)
}
\arguments{
\item{x}{A \code{workflow} object.}

\item{infra}{Should the core packages themselves be included in the result?
Those are workflows and parsnip (for the model), as well as recipes if the
workflow includes a recipes preprocessor and tailor if it includes a tailor
post-processor.}
}
\value{
A character vector.
}
\description{
Determine required packages for a workflow
}
\keyword{internal}
