% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-construction-methods.R
\name{rollback}
\alias{rollback}
\title{Revert execution state to previous target}
\usage{
rollback(x, ...)
}
\arguments{
\item{x}{Object of class "workflow" or "pipeline".}

\item{...}{Other arguments.}

\item{n}{Integer. If negative, roll back by that number of targets. If positive, go to the target number.}
}
\value{
A "workflow", or "pipeline" with altered "target" build states.
}
\description{
Sets targets as unbuilt, so running \code{execute} will re-build them.
}

