% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_of_modulo.R
\name{sum_of_modulo}
\alias{sum_of_modulo}
\title{Sum of a vectorized modulo function}
\usage{
sum_of_modulo(x, modulus)
}
\arguments{
\item{x}{Numeric vector}

\item{modulus}{Modulus as a 1L numeric vector.}
}
\description{
Sum of a vectorized modulo function
}
