% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows-package.R
\docType{package}
\name{workflows-package}
\alias{workflows}
\alias{workflows-package}
\title{workflows: Workflows}
\description{
Bundle a 'parsnip' model specification with a
    preprocessing formula or recipe for streamlined model fitting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/workflows}
  \item Report bugs at \url{https://github.com/DavisVaughan/workflows/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
