% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows-package.R
\docType{package}
\name{workflows-package}
\alias{workflows}
\alias{workflows-package}
\title{workflows: Modeling Workflows}
\description{
Managing both a 'parsnip' model and a preprocessor, such as a
    model formula or recipe from 'recipes', can often be challenging. The goal
    of 'workflows' is to streamline this process by bundling the model alongside
    the preprocessor, all within the same object.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidymodels/workflows}
  \item Report bugs at \url{https://github.com/tidymodels/workflows/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
