% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows-package.R
\docType{package}
\name{workflows-package}
\alias{workflows}
\alias{workflows-package}
\title{workflows: Modeling Workflows}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Managing both a 'parsnip' model and a preprocessor, such as a model formula or recipe from 'recipes', can often be challenging. The goal of 'workflows' is to streamline this process by bundling the model alongside the preprocessor, all within the same object.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidymodels/workflows}
  \item \url{https://workflows.tidymodels.org}
  \item Report bugs at \url{https://github.com/tidymodels/workflows/issues}
}

}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})

Authors:
\itemize{
  \item Davis Vaughan \email{davis@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC (03wc8by49) [copyright holder, funder]
}

}
\keyword{internal}
